Began rewrite. Also started refactoring the printing logic.
